Add Modular class, representing a modular arithmetic, fix misprints

+++ src/Numbers/Integer.cpp
@@ -18,7 +18,7 @@ private:
     
     size_t code;
 
-    IntegerError(int code) 
+    IntegerError(size_t code) 
         : code(code)
     {};
 };
